Rename rule `return-argument-limit` to `function-result-limit`
