+++ config.go
@@ -64,7 +64,7 @@ var allRules = append([]lint.Rule{
 	&rule.ConstantLogicalExprRule{},
 	&rule.BoolLiteralRule{},
 	&rule.RedefinesBuiltinIDRule{},
-	&rule.ReturnArgumentsLimitRule{},
+	&rule.FunctionResultsLimitRule{},
 }, defaultRules...)
 
 var allFormatters = []lint.Formatter{
